backend/requirements.txt
flask>=2.2
flask-socketio>=5.3
python-dotenv>=1.0
pymongo>=4.0
flask-cors>=3.0
eventlet>=0.33
werkzeug>=2.2
